// src/app/shared/interfaces/employee-interface.ts
export interface EmployeeInterface {
  id: bigint;
  first_name: string;
  last_name: string;
  employee_id: bigint;
  email: string;
  status:number;
  company_id: bigint;
  phone: bigint;
  street: string;
  city: string;
  state: string;
  postal_code: string;
  country: string;
  emergency_contact: string;
  emergency_phone: bigint;
}